+++ TextLayout/src/java/example/MainPanel.java
@@ -46,23 +46,23 @@ class TextLayoutPanel extends JComponent {
         float baseline = getHeight() / 2f;
 
         g2.setPaint(Color.RED);
-        g2.draw(new Line2D.Float(0, baseline, w, baseline));
+        g2.draw(new Line2D.Double(0, baseline, w, baseline));
 
         g2.setPaint(Color.GREEN);
         float ascent = baseline - TEXT_LAYOUT.getAscent();
-        g2.draw(new Line2D.Float(0, ascent, w, ascent));
+        g2.draw(new Line2D.Double(0, ascent, w, ascent));
 
         g2.setPaint(Color.BLUE);
         float descent = baseline + TEXT_LAYOUT.getDescent();
-        g2.draw(new Line2D.Float(0, descent, w, descent));
+        g2.draw(new Line2D.Double(0, descent, w, descent));
 
         g2.setPaint(Color.ORANGE);
         float leading = baseline + TEXT_LAYOUT.getDescent() + TEXT_LAYOUT.getLeading();
-        g2.draw(new Line2D.Float(0, leading, w, leading));
+        g2.draw(new Line2D.Double(0, leading, w, leading));
 
         g2.setPaint(Color.CYAN);
         float xheight = baseline - (float) TEXT_LAYOUT.getBlackBoxBounds(23, 24).getBounds().getHeight();
-        g2.draw(new Line2D.Float(0, xheight, w, xheight));
+        g2.draw(new Line2D.Double(0, xheight, w, xheight));
 
         g2.setPaint(Color.BLACK);
         TEXT_LAYOUT.draw(g2, 0f, baseline);
@@ -83,23 +83,23 @@ class GlyphVectorPanel extends JComponent {
         float baseline = getHeight() / 2f;
 
         g2.setPaint(Color.RED);
-        g2.draw(new Line2D.Float(0, baseline, w, baseline));
+        g2.draw(new Line2D.Double(0, baseline, w, baseline));
 
         g2.setPaint(Color.GREEN);
         float ascent = baseline - lm.getAscent();
-        g2.draw(new Line2D.Float(0, ascent, w, ascent));
+        g2.draw(new Line2D.Double(0, ascent, w, ascent));
 
         g2.setPaint(Color.BLUE);
         float descent = baseline + lm.getDescent();
-        g2.draw(new Line2D.Float(0, descent, w, descent));
+        g2.draw(new Line2D.Double(0, descent, w, descent));
 
         g2.setPaint(Color.ORANGE);
         float leading = baseline + lm.getDescent() + lm.getLeading();
-        g2.draw(new Line2D.Float(0, leading, w, leading));
+        g2.draw(new Line2D.Double(0, leading, w, leading));
 
         g2.setPaint(Color.CYAN);
         float xheight = baseline - (float) gv.getGlyphMetrics(23).getBounds2D().getHeight();
-        g2.draw(new Line2D.Float(0, xheight, w, xheight));
+        g2.draw(new Line2D.Double(0, xheight, w, xheight));
 
         g2.setPaint(Color.BLACK);
         g2.drawGlyphVector(gv, 0f, baseline);
